+++ server/sparkai-service/src/main/java/sparkai/service/entity/dataset/KnowledgeEmbeddingEntity.java
@@ -1,3 +1,12 @@
+// +----------------------------------------------------------------------
+// | SparkAI 基于大语言模型和 RAG 的知识库问答系统
+// +----------------------------------------------------------------------
+// | Copyright (c) 2022~2099 http://sparkai.sparkshop.cn All rights reserved.
+// +----------------------------------------------------------------------
+// | Licensed SparkAI 并不是自由软件，未经许可不能去掉 SparkAI 相关版权
+// +----------------------------------------------------------------------
+// | Author: NickBai  <1902822973@qq.com>
+// +----------------------------------------------------------------------
 package sparkai.service.entity.dataset;
 
 import com.baomidou.mybatisplus.annotation.IdType;
@@ -64,7 +73,7 @@ public class KnowledgeEmbeddingEntity implements Serializable {
     * 状态 1:正常 2:禁用
     */
     @TableField(value = "active")
-    private Short active;
+    private Integer active;
 
     /**
     * 创建时间
@@ -77,5 +86,4 @@ public class KnowledgeEmbeddingEntity implements Serializable {
     */
     @TableField(value = "update_time")
     private LocalDateTime updateTime;
-
 }
