+++ server/sparkai-service/src/main/java/sparkai/service/entity/dataset/KnowledgeParagraphEntity.java
@@ -1,3 +1,12 @@
+// +----------------------------------------------------------------------
+// | SparkAI 基于大语言模型和 RAG 的知识库问答系统
+// +----------------------------------------------------------------------
+// | Copyright (c) 2022~2099 http://sparkai.sparkshop.cn All rights reserved.
+// +----------------------------------------------------------------------
+// | Licensed SparkAI 并不是自由软件，未经许可不能去掉 SparkAI 相关版权
+// +----------------------------------------------------------------------
+// | Author: NickBai  <1902822973@qq.com>
+// +----------------------------------------------------------------------
 package sparkai.service.entity.dataset;
 
 import com.baomidou.mybatisplus.annotation.IdType;
@@ -58,13 +67,13 @@ public class KnowledgeParagraphEntity implements Serializable {
     * 状态 1:待索引 2:索引中 3:索引完成
     */
     @TableField(value = "status")
-    private Short status;
+    private Integer status;
 
     /**
     * 状态 1:正常 2:禁用
     */
     @TableField(value = "active")
-    private Short active;
+    private Integer active;
 
     /**
     * 状态数据
@@ -83,5 +92,4 @@ public class KnowledgeParagraphEntity implements Serializable {
     */
     @TableField(value = "update_time")
     private LocalDateTime updateTime;
-
 }
